move back to older implementation of db usage
