+++ src/postgresql/Dockerfile
@@ -2,11 +2,12 @@ FROM python:3.10-slim
 
 WORKDIR /app
 
-COPY ./postgresql/ /app/src/postgresql/
+COPY ./postgresql/ /app/postgresql/
 
-RUN pip install --no-cache-dir -r src/postgresql/requirements.txt
+RUN pip install --no-cache-dir -r postgresql/requirements.txt
 
-WORKDIR /app/src/postgresql/
+WORKDIR /app/postgresql/
+ENV PYTHONPATH=/app
 
 ENTRYPOINT ["alembic"]
 CMD ["upgrade", "head"]
